# api/swagger/models/api_my_sql_discover_request.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	strfmt "github.com/go-openapi/strfmt"

	"github.com/go-openapi/swag"
)

// APIMySQLDiscoverRequest api my SQL discover request
// swagger:model apiMySQLDiscoverRequest
type APIMySQLDiscoverRequest struct {

	// aws access key id
	AwsAccessKeyID string `json:"aws_access_key_id,omitempty"`

	// aws secret access key
	AwsSecretAccessKey string `json:"aws_secret_access_key,omitempty"`
}

// Validate validates this api my SQL discover request
func (m *APIMySQLDiscoverRequest) Validate(formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *APIMySQLDiscoverRequest) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *APIMySQLDiscoverRequest) UnmarshalBinary(b []byte) error {
	var res APIMySQLDiscoverRequest
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}