# api/swagger/mysql.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "mysql.proto",
    "version": "version not set"
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/v0/mysql/discover": {
      "post": {
        "operationId": "Discover",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/apiMySQLDiscoverResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apiMySQLDiscoverRequest"
            }
          }
        ],
        "tags": [
          "MySQL"
        ]
      }
    }
  },
  "definitions": {
...
    "apiMySQLDiscoverRequest": {
      "type": "object",
      "properties": {
        "aws_access_key_id": {
          "type": "string"
        },
        "aws_secret_access_key": {
          "type": "string"
        }
      }
    },
    "apiMySQLDiscoverResponse": {
      "type": "object",
      "properties": {
        "instances": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/apiMySQLInstance"
          }
        }
      }
    },
    "apiMySQLInstance": {
      "type": "object",
      "properties": {
        "node": {
          "$ref": "#/definitions/apiMySQLNode"
        },
        "service": {
          "$ref": "#/definitions/apiMySQLService"
        }
      }
    },
...
    "apiMySQLNode": {
      "type": "object",
      "properties": {
        "region": {
          "type": "string"
        },
        "name": {
          "type": "string"
        }
      }
    },
...
    "apiMySQLService": {
      "type": "object",
      "properties": {
        "address": {
          "type": "string"
        },
        "port": {
          "type": "integer",
          "format": "int64"
        },
        "engine": {
          "type": "string"
        },
        "engine_version": {
          "type": "string"
        }
      }
    }
  }
}